*: add jobspb.ScheduleID to distinguish ids
Previously, the IDs of job schedules we're just `int64s`. This was
identified as one of the causes of a recent bug where a schedule ID was
misinterpreted as a job ID.

This commit adds a wrapper type, `ScheduleID`, to clarify values which
correspond to the IDs of a job schedule vs the id of a job itself.

Fixes: #110909
Release note: None

+++ pkg/ccl/jobsccl/jobsprotectedtsccl/jobs_protected_ts_test.go
@@ -234,7 +234,7 @@ func testSchedulesProtectedTimestamp(
 			require.NoError(t, schedules.Create(ctx, sj))
 			deprecatedSpansToProtect := roachpb.Spans{{Key: keys.MinKey, EndKey: keys.MaxKey}}
 			targetToProtect := ptpb.MakeClusterTarget()
-			rec = jobsprotectedts.MakeRecord(uuid.MakeV4(), sj.ScheduleID(), ts,
+			rec = jobsprotectedts.MakeRecord(uuid.MakeV4(), int64(sj.ScheduleID()), ts,
 				deprecatedSpansToProtect, jobsprotectedts.Schedules, targetToProtect)
 			return ptp.WithTxn(txn).Protect(ctx, rec)
 		}))
